Remove EtcdCluster from Trillian app.yaml
The deployer does not wait for etcd-operator to install the CRD for
EtcdCluster before trying to use it, and so fails. Unfortunately, by
removing EtcdCluster from componentKinds, ownership information will not
be setup on the EtcdCluster and so it won't be uninstalled along with
Trillian.

https://github.com/GoogleCloudPlatform/marketplace-k8s-app-tools/issues/303
tracks a feature request for supporting Kubernetes operators that install
a CRD.