Check timeout should be less than check period
When adding an ingress controller to access the grafana service I received an error: 
```
i: Error 400: Invalid value for field 'resource.timeoutSec': '30'. TimeoutSec should be less than checkIntervalSec., invalid
```
the solution was to change the timeoutSeconds of the probe to be less than the period of the probe. This could go either way - increase the period to be longer than the timeout or reduce the timeout to be less than the period. I chose the latter.