Remove Stackdriver Metrics Exporter from nginx k8s click-to-deploy
Summary:
Cloud Monitoring / Stackdriver now recommends using [Managed Service for Prometheus](https://cloud.google.com/stackdriver/docs/managed-prometheus) to export Prometheus-format metrics to Cloud Monitoring instead of the old `prometheus-to-sd` sidecar. Why this is preferable:
- the metrics are of the form `prometheus.googleapis.com/...` instead of `custom.googleapis.com` or `kubernetes.io/nginx`. As such, they work OOTB in Grafana, OOTB Cloud Monitoring dashbaords and alerts. See the [integration setup guide](https://cloud.devsite.corp.google.com/stackdriver/docs/managed-prometheus/exporters/nginx).
- collection scales better because GMP is deployed as a daemonset (one pod per node) instead of one sidecar for every application pod
- cheaper and billed in samples instead of bytes, which is easier to understand

As a result, this deletes the older and not recommended prometheus-to-sd sidecar from the click-to-deploy configuration as an option.